refactor: #130 - 이미지 분석 관련 리펙토링
- Extension 및 Model 리펙토링
