+++ BestWish/Presentation/Extension/Int+Extension.swift
@@ -1,0 +1,17 @@
+//
+//  Int+Extension.swift
+//  BestWish
+//
+//  Created by Quarang on 6/26/25.
+//
+
+import Foundation
+
+extension Int {
+    func formattedPrice() -> String {
+        let formatter = NumberFormatter()
+        formatter.numberStyle = .decimal
+        formatter.groupingSeparator = ","
+        return formatter.string(from: NSNumber(value: self)) ?? "\(self)"
+    }
+}
